Bug fix when calling `dumpHands`
When calling `dumpHands`, `generatePose` is called without the
`float32Array` parameter. This leads to a runtime exception on line 128
when `float32Array`, which is undefined, is dereferenced to access the
`byteLength` field. This commit checks if `float32Array` is undefined
before the expression above is run.

Tested and works when calling `dumpHands`.